+++ Passwordless-dotnet.sln
@@ -3,15 +3,17 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Passwordless.Net.Tests", "tests\Passwordless.Net.Tests\Passwordless.Net.Tests.csproj", "{F64C850E-9923-43F1-BC84-432AFBBA4425}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Passwordless.Tests", "tests\Passwordless.Tests\Passwordless.Tests.csproj", "{F64C850E-9923-43F1-BC84-432AFBBA4425}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Passwordless.Net", "src\Passwordless.Net\Passwordless.Net.csproj", "{A01503A8-6AB9-43A7-AC5A-4EAE091B07B6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Passwordless", "src\Passwordless\Passwordless.csproj", "{A01503A8-6AB9-43A7-AC5A-4EAE091B07B6}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{F239A76C-408E-4919-AB70-4499425E6F29}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 		.gitignore = .gitignore
 		Directory.Build.props = Directory.Build.props
+		README.md = README.md
+		LICENSE = LICENSE
 	EndProjectSection
 EndProject
 Global
