--- src/Passwordless.Net/Models/AliasPointer.cs
@@ -1,15 +1,0 @@
-﻿namespace Passwordless.Net;
-
-public class AliasPointer
-{
-    public AliasPointer(string userId, string alias, string plaintext)
-    {
-        UserId = userId;
-        Alias = alias;
-        Plaintext = plaintext;
-    }
-
-    public string UserId { get; }
-    public string Alias { get; }
-    public string Plaintext { get; }
-}
