--- src/Passwordless.Net/Models/DeleteCredentialRequest.cs
@@ -1,11 +1,0 @@
-namespace Passwordless.Net.Models;
-
-internal class DeleteCredentialRequest
-{
-    public DeleteCredentialRequest(string credentialId)
-    {
-        CredentialId = credentialId;
-    }
-
-    public string CredentialId { get; }
-}
