--- src/Passwordless.Net/Models/DeleteUserRequest.cs
@@ -1,11 +1,0 @@
-namespace Passwordless.Net.Models;
-
-internal class DeleteUserRequest
-{
-    public DeleteUserRequest(string userId)
-    {
-        UserId = userId;
-    }
-
-    public string UserId { get; }
-}
