--- src/Passwordless.Net/Models/PasswordlessUserSummary.cs
@@ -1,20 +1,0 @@
-namespace Passwordless.Net;
-
-public class PasswordlessUserSummary
-{
-    public PasswordlessUserSummary(string userId, IReadOnlyList<string> aliases, int credentialsCount,
-        int aliasCount, DateTime? lastUsedAt)
-    {
-        UserId = userId;
-        Aliases = aliases;
-        CredentialsCount = credentialsCount;
-        AliasCount = aliasCount;
-        LastUsedAt = lastUsedAt;
-    }
-
-    public string UserId { get; }
-    public IReadOnlyList<string> Aliases { get; }
-    public int CredentialsCount { get; }
-    public int AliasCount { get; }
-    public DateTime? LastUsedAt { get; }
-}
