--- src/Passwordless.Net/Models/RegisterTokenResponse.cs
@@ -1,13 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Passwordless.Net.Models;
-
-public class RegisterTokenResponse
-{
-    public RegisterTokenResponse(string token)
-    {
-        Token = token;
-    }
-
-    public string Token { get; }
-}
