--- src/Passwordless.Net/Models/UsersCount.cs
@@ -1,11 +1,0 @@
-namespace Passwordless.Net;
-
-public class UsersCount
-{
-    public UsersCount(int count)
-    {
-        Count = count;
-    }
-
-    public int Count { get; }
-}
