--- src/Passwordless.Net/Models/VerifyTokenRequest.cs
@@ -1,11 +1,0 @@
-namespace Passwordless.Net.Models;
-
-internal class VerifyTokenRequest
-{
-    public VerifyTokenRequest(string token)
-    {
-        Token = token;
-    }
-
-    public string Token { get; }
-}
