--- src/Passwordless.Net/PasswordlessExtensions.cs
@@ -1,9 +1,0 @@
-﻿namespace Passwordless.Net;
-
-public static class PasswordlessExtensions
-{
-    public static string ToBase64Url(this byte[] bytes)
-    {
-        return Base64Url.Encode(bytes);
-    }
-}
