+++ src/Passwordless/Models/AliasPointer.cs
@@ -1,0 +1,15 @@
+﻿namespace Passwordless;
+
+public class AliasPointer
+{
+    public AliasPointer(string userId, string alias, string plaintext)
+    {
+        UserId = userId;
+        Alias = alias;
+        Plaintext = plaintext;
+    }
+
+    public string UserId { get; }
+    public string Alias { get; }
+    public string Plaintext { get; }
+}
