+++ src/Passwordless/Models/DeleteCredentialRequest.cs
@@ -1,0 +1,11 @@
+namespace Passwordless.Models;
+
+internal class DeleteCredentialRequest
+{
+    public DeleteCredentialRequest(string credentialId)
+    {
+        CredentialId = credentialId;
+    }
+
+    public string CredentialId { get; }
+}
