+++ src/Passwordless/Models/DeleteUserRequest.cs
@@ -1,0 +1,11 @@
+namespace Passwordless.Models;
+
+internal class DeleteUserRequest
+{
+    public DeleteUserRequest(string userId)
+    {
+        UserId = userId;
+    }
+
+    public string UserId { get; }
+}
