+++ src/Passwordless/Models/ListResponse.cs
@@ -1,0 +1,11 @@
+namespace Passwordless.Models;
+
+internal class ListResponse<T>
+{
+    public ListResponse(IReadOnlyList<T> values)
+    {
+        Values = values;
+    }
+
+    public IReadOnlyList<T> Values { get; }
+}
