+++ src/Passwordless/Models/RegisterTokenResponse.cs
@@ -1,0 +1,13 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Passwordless.Models;
+
+public class RegisterTokenResponse
+{
+    public RegisterTokenResponse(string token)
+    {
+        Token = token;
+    }
+
+    public string Token { get; }
+}
