+++ src/Passwordless/Models/UsersCount.cs
@@ -1,0 +1,11 @@
+namespace Passwordless;
+
+public class UsersCount
+{
+    public UsersCount(int count)
+    {
+        Count = count;
+    }
+
+    public int Count { get; }
+}
