+++ src/Passwordless/Models/VerifyTokenRequest.cs
@@ -1,0 +1,11 @@
+namespace Passwordless.Models;
+
+internal class VerifyTokenRequest
+{
+    public VerifyTokenRequest(string token)
+    {
+        Token = token;
+    }
+
+    public string Token { get; }
+}
