+++ src/Passwordless/Passwordless.csproj
@@ -1,0 +1,46 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>net462;net6.0;net7.0;netstandard2.0</TargetFrameworks>
+    <TargetFrameworks Condition="'$(IncludePreview)' == 'true'">$(TargetFrameworks);$(CurrentPreviewTfm)</TargetFrameworks>
+    <PublishAot Condition="'$(TargetFrameworkIdentifier)' == '.NETCoreApp' AND $([MSBuild]::VersionGreaterThanOrEquals($(TargetFrameworkVersion), '8.0'))">true</PublishAot>
+  </PropertyGroup>
+
+  <!-- Packaging-related properties -->
+  <PropertyGroup>
+    <Authors>Bitwarden</Authors>
+    <RepositoryUrl>https://github.com/passwordless/passwordless-dotnet</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+  </PropertyGroup>
+
+  <!-- .NET 6 first introduced these types, for all others we include our own pollyfill so that it builds -->
+  <PropertyGroup Condition="'$(TargetFrameworkIdentifier)' != '.NETCoreApp' OR $([MSBuild]::VersionLessThan($(TargetFrameworkVersion), '6.0'))">
+    <DefineConstants>$(DefineConstants);INCLUDE_DYNAMICALLY_ACCESSED_MEMBERS</DefineConstants>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Http" Version="7.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Options" Version="7.0.1" />
+  </ItemGroup>
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'net462'">
+    <PackageReference Include="System.Text.Json" Version="7.0.3" />
+    <PackageReference Include="System.Net.Http.Json" Version="7.0.1" />
+    <PackageReference Include="System.Memory" Version="4.5.5" />
+  </ItemGroup>
+
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
+    <PackageReference Include="System.Text.Json" Version="7.0.3" />
+    <PackageReference Include="System.Net.Http.Json" Version="7.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="../../README.md" Pack="true" PackagePath="/" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <InternalsVisibleTo Include="Passwordless.AdminConsole" />
+  </ItemGroup>
+
+</Project>
