+++ src/Passwordless/PasswordlessDelegatingHandler.cs
@@ -1,0 +1,32 @@
+using System.Net.Http;
+using System.Net.Http.Json;
+using Passwordless.Helpers;
+
+namespace Passwordless;
+
+internal class PasswordlessDelegatingHandler : DelegatingHandler
+{
+    protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+    {
+        var response = await base.SendAsync(request, cancellationToken);
+
+        if (request.ShouldSkipErrorHandling())
+        {
+            return response;
+        }
+
+        if (!response.IsSuccessStatusCode
+            && string.Equals(response.Content.Headers.ContentType?.MediaType, "application/problem+json", StringComparison.OrdinalIgnoreCase))
+        {
+            // Attempt to read problem details
+            var problemDetails = await response.Content.ReadFromJsonAsync(
+                PasswordlessSerializerContext.Default.PasswordlessProblemDetails,
+                cancellationToken);
+
+            // Throw exception
+            throw new PasswordlessApiException(problemDetails!);
+        }
+
+        return response;
+    }
+}
