+++ src/Passwordless/PasswordlessExtensions.cs
@@ -1,0 +1,9 @@
+﻿namespace Passwordless;
+
+public static class PasswordlessExtensions
+{
+    public static string ToBase64Url(this byte[] bytes)
+    {
+        return Base64Url.Encode(bytes);
+    }
+}
