--- tests/Passwordless.Net.Tests/ApiFactAttribute.cs
@@ -1,14 +1,0 @@
-// Uncomment line while running API in mock mode to run tests
-// #define RUNNING_API
-
-namespace Passwordless.Net.Tests;
-
-public class ApiFactAttribute : FactAttribute
-{
-    public ApiFactAttribute()
-    {
-#if !RUNNING_API
-        Skip = "These tests are skipped unless you are running the API locally.";
-#endif
-    }
-}
